requirements.txt
numpy
keras
Pillow
scikit-learn